feat:#117 리뷰순 버튼을 누르면 리뷰순으로 보여주기
