+++ src/components/Detail/detailReview/ReviewFilterButton.tsx
@@ -4,7 +4,12 @@ import Button from '@/components/common/Button/Button';
 import { theme } from '@/styles/theme';
 import * as S from '../Detail.styles';
 
-const ReviewFilterButton = () => {
+type FilterProps = {
+  // handleArrangement: (index: number) => void;
+  byRating: () => void;
+};
+
+const ReviewFilterButton = ({ byRating }: FilterProps) => {
   const [filter, setFilter] = useState([
     {
       title: '최신순',
@@ -31,6 +36,9 @@ const ReviewFilterButton = () => {
       }
     });
     setFilter(updatedFilter);
+    // handleArrangement(index);
+
+    if (index === 1) byRating();
   };
 
   return (
