+++ src/pages/DetailPage/DetailPage.tsx
@@ -3,13 +3,9 @@ import { useParams } from 'react-router-dom';
 import styled from '@emotion/styled';
 
 import { CartItemAPI, getProduct } from '@/apis/product';
-import { getReview } from '@/apis/review';
 import DetailCategory from '@/components/Detail/detailCategory/DetailCategory';
 import InformationBox from '@/components/Detail/detailInformation/InformationBox';
-import ReviewBox from '@/components/Detail/detailReview/ReviewBox';
-import ReviewButton from '@/components/Detail/detailReview/ReviewButton';
-import ReviewFilterButton from '@/components/Detail/detailReview/ReviewFilterButton';
-import ReviewWrite from '@/components/Detail/detailReview/ReviewWrite';
+import Review from '@/components/Detail/detailReview/Review';
 import DetailFooter from '@/components/DetailFooter/DetailFooter';
 import DetailHeader from '@/components/DetailHeader/DetailHeader';
 import ProductOption from '@/components/DetailHeader/ProductOption';
@@ -30,37 +26,13 @@ export type DetailProduct = {
   }[];
 };
 
-export type DetailReview = {
-  productId: number;
-  productName: string;
-  options: string;
-  reviewId: number;
-  author: string;
-  title: string;
-  content: string;
-  starPoint: number;
-  imageUrl: string;
-  createdAt: string;
-};
-
 const DetailPage = () => {
   const { productId } = useParams();
   const [product, setProduct] = useState<DetailProduct>();
-  const [review, setReview] = useState<DetailReview[]>([]);
   const [cartProduct, setCartProduct] = useState<CartItemAPI[]>([]);
-  const [isReview, setIsReview] = useState<boolean>(false);
   const [prevCategory, setPrevCategory] = useState<string>('');
 
-  useEffect(() => {
-    if (isReview) {
-      getReview(Number(productId)).then((result) => {
-        if (result.status === 200) {
-          const data = result.data;
-          setReview((prevReview) => [...prevReview, ...data]);
-        }
-      });
-    }
-  }, [isReview, productId]);
+  const [isReview, setIsReview] = useState<boolean>(false);
 
   useEffect(() => {
     getProduct(Number(productId)).then((result) => {
@@ -125,12 +97,7 @@ const DetailPage = () => {
         <InformationBox />
       ) : (
         <>
-          <ReviewButton />
-          <ReviewWrite />
-          <ReviewFilterButton />
-          {review?.map((item, idx) => {
-            return <ReviewBox key={idx} review={item} />;
-          })}
+          <Review productId={Number(productId)} isReview={isReview} />
         </>
       )}
       <DetailFooter cartProduct={cartProduct} />
